Atari 2600: Place all TIA logic in a single class
